+++ src/lib/output/themes/default/DefaultThemeRenderContext.ts
@@ -1,11 +1,11 @@
 import type { RendererHooks } from "../..";
-import type {
+import {
+    Comment,
     CommentDisplayPart,
     ReferenceType,
     Reflection,
 } from "../../../models";
 import type { NeverIfInternal, Options } from "../../../utils";
-import { displayPartsToMarkdown } from "../lib";
 import type { DefaultTheme } from "./DefaultTheme";
 import { defaultLayout } from "./layouts/default";
 import { index } from "./partials";
@@ -62,14 +62,14 @@ export class DefaultThemeRenderContext {
         return url ? this.theme.markedPlugin.getRelativeUrl(url) : url;
     };
 
-    urlTo = (reflection: Reflection) => this.relativeURL(reflection.url);
+    urlTo = (reflection: Reflection) => this.relativeURL(reflection.url)!;
 
     markdown = (
         md: readonly CommentDisplayPart[] | NeverIfInternal<string | undefined>
     ) => {
         if (md instanceof Array) {
             return this.theme.markedPlugin.parseMarkdown(
-                displayPartsToMarkdown(md, this.urlTo)
+                Comment.displayPartsToMarkdown(md, this.urlTo)
             );
         }
         return md ? this.theme.markedPlugin.parseMarkdown(md) : "";